+++ app/src/main/java/com/turtlecoin/turtlewallet/EditContactActivity.kt
@@ -5,7 +5,6 @@ import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
 import android.view.View
-import android.widget.Button
 import android.widget.Toast
 import com.google.zxing.integration.android.IntentIntegrator
 import com.turtlecoin.turtlewallet.util.AddressValidator
